refactor(scripts): use inheritance for cleaning/merging scripts
