+++ django/common/scripts/__init__.py
@@ -1,3 +1,3 @@
 # flake8: noqa
 from . import utils
-from .scripts_repository import Script, ScriptsRepository
+from .scripts_repository import CleaningScript, MergingScript, Script, ScriptNotFound, ScriptsRepository
