+++ django/common/scripts/scripts_repository.py
@@ -1,9 +1,16 @@
+import logging
 from collections import OrderedDict
 from dataclasses import dataclass
 from inspect import getdoc, getmembers, isfunction
-from typing import Callable
+from typing import Callable, List
 
 from common.scripts import cleaning, merging
+
+logger = logging.getLogger(__name__)
+
+
+class ScriptNotFound(Exception):
+    pass
 
 
 @dataclass(frozen=True)
@@ -12,6 +19,33 @@ class Script:
     name: str
     description: str
     category: str
+
+    def __eq__(self, operand) -> bool:
+        return self.name == operand.name
+
+    def apply(self, data_columns, static_inputs: List[str], attr_path, primary_key):
+        raise NotImplementedError
+
+
+class MergingScript(Script):
+    def apply(self, data_columns, static_inputs: List[str], attr_path, primary_key):
+        try:
+            args = data_columns + static_inputs
+            return self.func(*args)
+        except Exception as e:
+            logger.exception(
+                f"{self.name}: Error merging columns for attribute at path " f"{attr_path} (at id={primary_key}): {e}"
+            )
+            return data_columns
+
+
+class CleaningScript(Script):
+    def apply(self, data_column, col_name, primary_key):
+        try:
+            return [self.func(val) for val in data_column]
+        except Exception as e:
+            logger.exception(f"{self.script.name}: Error cleaning {col_name} (at id = {primary_key}): {e}")
+            return data_column
 
 
 class ScriptsRepository:
@@ -24,12 +58,12 @@ class ScriptsRepository:
         self.scripts = OrderedDict()
 
         for script_name, script in getmembers(cleaning, isfunction):
-            self.scripts[script_name] = Script(
+            self.scripts[script_name] = CleaningScript(
                 func=script, name=script_name, description=getdoc(script), category="cleaning"
             )
 
         for script_name, script in getmembers(merging, isfunction):
-            self.scripts[script_name] = Script(
+            self.scripts[script_name] = MergingScript(
                 func=script, name=script_name, description=getdoc(script), category="merging"
             )
 
@@ -38,4 +72,4 @@ class ScriptsRepository:
         try:
             return self.scripts[name]
         except KeyError:
-            raise NameError(f"Script {name} not found.")
+            raise ScriptNotFound(f"Script {name} not found.")
