+++ tests/river/unit/transformer/transform/test_dataframe.py
@@ -1,12 +1,10 @@
 from unittest import mock
 
-from common.scripts import Script
+from common.scripts import CleaningScript, MergingScript
 from river.common.analyzer.attribute import Attribute
-from river.common.analyzer.cleaning_script import CleaningScript
 from river.common.analyzer.concept_map import ConceptMap
 from river.common.analyzer.condition import CONDITION_FLAG, Condition
 from river.common.analyzer.input_group import InputGroup
-from river.common.analyzer.merging_script import MergingScript
 from river.common.analyzer.sql_column import SqlColumn
 from river.transformer import dataframe
 
@@ -14,7 +12,7 @@ from river.transformer import dataframe
 @mock.patch("river.common.analyzer.sql_column.hashlib.sha1")
 def test_clean_data(mock_sha1, dict_map_gender, dict_map_code):
     cleaning_script = CleaningScript(
-        Script(name="clean", func=lambda arg: arg.replace("dirty", ""), description=None, category="cleaning")
+        name="clean", func=lambda arg: arg.replace("dirty", ""), description=None, category="cleaning"
     )
     mock_sha1.return_value.hexdigest.return_value = "hash"
 
@@ -94,9 +92,7 @@ def test_clean_data(mock_sha1, dict_map_gender, dict_map_code):
 
 def test_merge_by_attributes():
     merging_script = MergingScript(
-        Script(
-            name="merge", func=lambda *args: "".join(arg for arg in args if arg), description=None, category="merging"
-        )
+        name="merge", func=lambda *args: "".join(arg for arg in args if arg), description=None, category="merging"
     )
     attr_name = Attribute("name")
     group = InputGroup(id_="id_name", attribute=attr_name, columns=[SqlColumn("PUBLIC", "PATIENTS", "NAME")])
